requirements.txt
fastapi
uvicorn[standard]
websockets
wsproto
jinja2
python-multipart
pillow
tqdm
fairscale==0.4.13
open_clip_torch
timm>=1.0.0
transformers>=5.5.0,<6
sentencepiece
protobuf
hf_xet
insightface>=0.7.3
opencv-python-headless>=4.9.0
git+https://github.com/openai/CLIP.git
llama-cpp-python>=0.3.7
faiss-cpu
